split worker and worker pool in separated file. code cleanup.
